pyproject.toml
[tool.poetry]
name = "lupin"
version = "0.1.0"
description = ""
authors = ["cardosofede <federico.cardoso.e@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["job-queue"], version = "^20.7"}
pandas = "^2.1.3"
langchain = "^0.0.346"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.6.0"
black = "^23.12.0"
isort = "^5.13.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
